+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Überzug++ as fallback for image rendering.
 - Add Unicode half-blocks image rendering as fallback.
 - Add a command for copying links to clipboard.
+- Add chafa as fallback for image rendering.
 
 ### Fixed
 - Handle cell sizes with floating point numbers.